remove old code + move sampling to planning
