+++ hat/assets/js/apps/Iaso/domains/plannings/hooks/requests/useBulkDeleteAssignments.tsx
