+++ hat/assets/js/apps/Iaso/domains/plannings/sampling/OpenhexaIntegrationDrawer.tsx
