+++ hat/assets/js/apps/Iaso/domains/plannings/sampling/components/PipelineInfos.tsx
