+++ hat/assets/js/apps/Iaso/domains/plannings/sampling/customForms/ExcludedOrgUnits.tsx
@@ -1,6 +1,6 @@
 import React, { FunctionComponent, useCallback, useMemo } from 'react';
 import InputComponent from 'Iaso/components/forms/InputComponent';
-import { useGetOrgUnitsByOrgUnitTypeId } from 'Iaso/domains/assignments/hooks/requests/useGetOrgUnits';
+import { useGetOrgUnitsByOrgUnitTypeId } from 'Iaso/domains/plannings/hooks/requests/useGetOrgUnits';
 import { Planning } from 'Iaso/domains/plannings/types';
 import MESSAGES from '../../messages';
 import { ParameterValues } from './LQASForm';
