+++ hat/assets/js/apps/Iaso/domains/plannings/sampling/customForms/constants.ts
